+++ src/dynamic_vertex.rs
@@ -7,15 +7,7 @@ use itertools::Itertools;
 #[cfg(any(test, feature = "serde"))]
 use serde::{Deserialize, Serialize};
 
-use crate::{
-    point::{Point, HasCenter},
-    corner::Corner,
-    flip_axes::FlipAxes,
-    prelude::DynamicTile,
-    quarter_turns::QuarterTurns,
-    tile::Tile,
-    vector::Vector,
-};
+use crate::prelude::*;
 
 /// A vertex in a dynamically sized 2d space
 #[must_use]
@@ -92,8 +84,8 @@ impl DynamicVertex {
     /// Get the nearest vertex to this center.
     /// Will round away from 0.0
     #[must_use]
-    #[cfg(any(test, feature = "std"))]
-    pub fn from_center(center: &Point, scale: f32) -> Self {
+    #[cfg(any(test, all(feature = "std", feature="glam")))]
+    pub fn from_center(center: &glam::f32::Vec2, scale: f32) -> Self {
         let x = center.x / scale;
         let y = center.y / scale;
 
@@ -105,13 +97,14 @@ impl DynamicVertex {
     }
 }
 
+#[cfg(any(test, feature = "glam"))]
 impl HasCenter for DynamicVertex {
     #[must_use]
-    fn get_center(&self, scale: f32) -> Point {
+    fn get_center(&self, scale: f32) -> glam::f32::Vec2 {
         let x = scale * (self.0.x as f32);
         let y = scale * (self.0.y as f32);
 
-        Point { x, y }
+        glam::f32::Vec2 { x, y }
     }
 }
 
@@ -152,7 +145,7 @@ mod tests {
     pub fn test_center() {
         let vertex: DynamicVertex = Vector::new(-2, 3).into();
 
-        assert_eq!(vertex.get_center(3.0), Point::new(-6.0, 9.0))
+        assert_eq!(vertex.get_center(3.0), glam::f32::Vec2::new(-6.0, 9.0))
     }
     #[test]
     pub fn test_add() {
@@ -186,7 +179,7 @@ mod tests {
     #[test]
     pub fn test_from_center(){
         fn t(x: f32, y: f32, scale: f32, expected_x: i8, expected_y: i8){
-            let actual  = DynamicVertex::from_center(&Point { x, y }, scale);
+            let actual  = DynamicVertex::from_center(&glam::f32::Vec2 { x, y }, scale);
             assert_eq!(DynamicVertex(Vector { x: expected_x, y: expected_y }), actual)
         }